rescope to Lich::Resources
